Some but not all of Michael Vines' style changes in here. Still WIP, brought in borsh and combined files.
