+++ metadata/program/Cargo.toml
@@ -17,16 +17,17 @@ test-dump-genesis-accounts = []
 arrayref = "0.3.6"
 num-derive = "0.3"
 num-traits = "0.2"
-solana-program = "1.5.13"
+solana-program = "1.6.2"
 spl-token = { path = "../../token/program", features = [ "no-entrypoint" ] }
 thiserror = "1.0"
+borsh = "0.8.2"
 
 [dev-dependencies]
 assert_matches = "1.4.0"
 log = "0.4.8"
 proptest = "0.10"
-solana-program-test = "1.5.13"
-solana-sdk = "1.5.13"
+solana-program-test = "1.6.2"
+solana-sdk = "1.6.2"
 serde = "1.0"
 serde_yaml = "0.8"
 tokio = { version = "0.3", features = ["macros"]}
