--- metadata/program/src/state/mod.rs
@@ -1,9 +1,0 @@
-const UNINITIALIZED_VERSION: u8 = 0;
-
-///prefix
-pub const PREFIX: &str = "metadata";
-
-/// Metadata
-pub mod metadata;
-///  Owner
-pub mod owner;
