+++ metadata/program/src/utils.rs
@@ -19,18 +19,6 @@ pub fn assert_rent_exempt(rent: &Rent, account_info: &AccountInfo) -> ProgramRes
     }
 }
 
-/// assert ununitialized account
-pub fn assert_uninitialized<T: Pack + IsInitialized>(
-    account_info: &AccountInfo,
-) -> Result<T, ProgramError> {
-    let account: T = T::unpack_unchecked(&account_info.data.borrow())?;
-    if account.is_initialized() {
-        Err(MetadataError::AlreadyInitialized.into())
-    } else {
-        Ok(account)
-    }
-}
-
 /// assert initialized account
 pub fn assert_initialized<T: Pack + IsInitialized>(
     account_info: &AccountInfo,
@@ -46,14 +34,14 @@ pub fn assert_initialized<T: Pack + IsInitialized>(
 /// Create account from scratch, stolen from Wormhole, slightly altered for my purposes
 /// https://github.com/bartosz-lipinski/wormhole/blob/8478735ea7525043635524a62db2751e59d2bc38/solana/bridge/src/processor.rs#L1335
 #[inline(always)]
-pub fn create_account_raw<T: Pack>(
+pub fn create_account_raw<T>(
     accounts: &[AccountInfo],
     new_account: &Pubkey,
     payer: &Pubkey,
     owner: &Pubkey,
     seeds: &[&[u8]],
+    size: u64,
 ) -> Result<(), ProgramError> {
-    let size = T::LEN;
     let ix = create_account(
         payer,
         new_account,
