+++ metadata/test/Cargo.toml
@@ -9,9 +9,10 @@ edition = "2018"
 publish = false
 
 [dependencies]
-solana-client = "1.5.6"
-solana-program = "1.5.6"
-solana-sdk = "1.5.6"
+solana-client = "1.6.2"
+solana-program = "1.6.2"
+solana-sdk = "1.6.2"
 bincode = "1.3.2"
+borsh = "0.8.2"
 spl-metadata = { path = "../program", features = [ "no-entrypoint" ] }
 spl-token = { path = "../../token/program", features = [ "no-entrypoint" ] }
