+++ metadata/test/src/main.rs
@@ -1,5 +1,5 @@
 use solana_client::rpc_client::RpcClient;
-use solana_program::program_pack::Pack;
+use solana_program::{borsh::try_from_slice_unchecked, program_pack::Pack};
 use solana_sdk::{
     pubkey::Pubkey,
     signature::{read_keypair_file, Signer},
@@ -8,18 +8,18 @@ use solana_sdk::{
 };
 use spl_metadata::{
     instruction::{create_metadata_accounts, init_metadata_accounts, update_metadata_accounts},
-    state::metadata::{Metadata, NAME_LENGTH, SYMBOL_LENGTH},
-    state::PREFIX,
+    state::{Metadata, PREFIX},
 };
+
 use spl_token::{instruction::initialize_mint, state::Mint};
 use std::str::FromStr;
 // -------- UPDATE START -------
 
-const KEYPAIR_PATH: &str = "/your/path";
-const METADATA_PROGRAM_PUBKEY: &str = "metaTMm9vFU2h7m97KYJA9k2JQUG21rfb434rWUQBiH";
-const NEW_MINT_PATH: &str = "/your/path";
+const KEYPAIR_PATH: &str = "/Users/jprince/.config/solana/id.json";
+const METADATA_PROGRAM_PUBKEY: &str = "Xk2Cihp7vnSWdcfDjH6W3H79WoFfADyqQQW6cENNXDS";
+const NEW_MINT_PATH: &str = "/Users/jprince/.config/solana/mint.json";
 const TOKEN_PROGRAM_PUBKEY: &str = "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA";
-const CLUSTER_ADDRESS: &str = "https://api.mainnet-beta.solana.com";
+const CLUSTER_ADDRESS: &str = "https://devnet.solana.com";
 
 // -------- UPDATE END ---------
 pub fn main() {
@@ -36,23 +36,13 @@ pub fn main() {
         new_mint_key.as_ref(),
     ];
     let (metadata_key, _) = Pubkey::find_program_address(metadata_seeds, &program_key);
-    let mut name: [u8; NAME_LENGTH] = [0; NAME_LENGTH];
-    let mut symbol: [u8; SYMBOL_LENGTH] = [0; SYMBOL_LENGTH];
 
-    let name_bytes = "Billy1".as_bytes();
-    for n in 0..(NAME_LENGTH - 1) {
-        if n < name_bytes.len() {
-            name[n] = name_bytes[n];
-        }
-    }
-
-    let symbol_bytes = "Bob1".as_bytes();
-    for n in 0..(SYMBOL_LENGTH - 1) {
-        if n < symbol_bytes.len() {
-            symbol[n] = symbol_bytes[n];
-        }
-    }
-    let owner_seeds = &[PREFIX.as_bytes(), &program_key.as_ref(), &name, &symbol];
+    let owner_seeds = &[
+        PREFIX.as_bytes(),
+        &program_key.as_ref(),
+        &"Billy3".as_bytes(),
+        &"Bob3".as_bytes(),
+    ];
     let (owner_key, _) = Pubkey::find_program_address(owner_seeds, &program_key);
 
     let mut transaction = Transaction::new_with_payer(
@@ -81,8 +71,8 @@ pub fn main() {
                 new_mint.pubkey(),
                 payer.pubkey(),
                 payer.pubkey(),
-                "Billy1",
-                "Bob1",
+                "Billy3".to_owned(),
+                "Bob3".to_owned(),
             ),
             init_metadata_accounts(
                 program_key,
@@ -91,16 +81,16 @@ pub fn main() {
                 new_mint.pubkey(),
                 payer.pubkey(),
                 payer.pubkey(),
-                "Billy1",
-                "Bob1",
-                "www.billybob.com",
+                "Billy3".to_owned(),
+                "Bob3".to_owned(),
+                "www.billybob.com".to_owned(),
             ),
             update_metadata_accounts(
                 program_key,
                 metadata_key,
                 owner_key,
                 payer.pubkey(),
-                "www.aol.com",
+                "www.aol.com".to_owned(),
             ),
         ],
         Some(&payer.pubkey()),
@@ -109,9 +99,9 @@ pub fn main() {
     transaction.sign(&[&payer, &new_mint], recent_blockhash);
     client.send_and_confirm_transaction(&transaction).unwrap();
     let account = client.get_account(&metadata_key).unwrap();
-    let metadata = Metadata::unpack(&account.data).unwrap();
+    let metadata: Metadata = try_from_slice_unchecked(&account.data).unwrap();
     println!(
         "If this worked correctly, updated metadata should have aol: {:?} ",
-        std::str::from_utf8(&metadata.uri).unwrap()
+        metadata.uri
     );
 }
